feat(sfx): move SFX commands to a dedicated page
Extracts the SFX table from the dashboard into its own /sfx route,
view, and client script. Updates nav tab order to: Dashboard, SFX,
Commands, Counters, Command Monitor, Streams, Users. Bumps service
worker cache to v7 to pre-cache the new sfx.js asset.
