+++ views/partials/nav.ejs
@@ -4,12 +4,13 @@
   <div class="navbar-collapse" id="navbar-collapse" aria-hidden="false">
     <div class="navbar-menu">
       <a href="/" class="nav-item">Dashboard</a>
+      <a href="/sfx" class="nav-item">SFX</a>
       <% if (user && user.accessLevel >= 2) { %>
-      <a href="/admin/users" class="nav-item">Users</a>
       <a href="/admin/commands" class="nav-item">Commands</a>
-      <a href="/admin/command-monitor" class="nav-item">Command Monitor</a>
       <a href="/admin/counters" class="nav-item">Counters</a>
+      <a href="/admin/command-monitor" class="nav-item">Command Monitor</a>
       <a href="/admin/streams" class="nav-item">Streams</a>
+      <a href="/admin/users" class="nav-item">Users</a>
       <% } %>
     </div>
     <div class="navbar-end">
