+++ views/sfx.ejs
@@ -1,0 +1,84 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>BCUK Bot — SFX</title>
+  <link rel="stylesheet" href="/style.css">
+  <%- include('partials/pwa-head') %>
+</head>
+<body data-csrf-token="<%= csrfToken %>">
+  <%- include('partials/nav') %>
+
+  <main class="container">
+
+    <!-- ── SFX Commands ── -->
+    <section class="section">
+      <div class="section-header">
+        <h2 class="section-title">SFX Commands (<span id="cmd-count"><%= triggers.length %></span>)</h2>
+        <input type="search" id="sfx-search" class="search-input" placeholder="Filter commands…">
+      </div>
+
+      <div class="card">
+        <div class="table-scroll">
+        <table class="table" id="sfx-table" aria-label="SFX commands">
+          <thead>
+            <tr>
+              <th scope="col">Command</th>
+              <th scope="col">Category</th>
+              <th scope="col">Files</th>
+              <th scope="col">Status</th>
+              <th scope="col"></th>
+            </tr>
+          </thead>
+          <tbody>
+            <% triggers.forEach(function(t) { %>
+            <tr class="sfx-row<%= t.hidden ? ' row-hidden' : '' %>" data-command="<%= t.triggerCommand.toLowerCase() %>" data-category="<%= (t.categoryName || '').toLowerCase() %>">
+              <td class="mono"><%= t.triggerCommand %></td>
+              <td><% if (t.categoryName) { %><%= t.categoryName %><% } else { %><em class="muted">None</em><% } %></td>
+              <td><%= t.files.length %> file<%= t.files.length !== 1 ? 's' : '' %></td>
+              <td>
+                <% if (t.hidden) { %>
+                <span class="badge badge-hidden">Hidden</span>
+                <% } else { %>
+                <span class="badge badge-active">Active</span>
+                <% } %>
+              </td>
+              <td>
+                <% if (t.files.length > 0) { %>
+                <button type="button" class="btn btn-xs btn-toggle-files">▶ Files</button>
+                <% } %>
+              </td>
+            </tr>
+            <% if (t.files.length > 0) { %>
+            <tr class="files-row is-hidden">
+              <td colspan="5">
+                <table class="inner-table">
+                  <thead><tr><th scope="col">File</th><th scope="col">Weight</th><th scope="col">Status</th></tr></thead>
+                  <tbody>
+                    <% t.files.forEach(function(f) { %>
+                    <tr>
+                      <td class="mono"><%= f.file %></td>
+                      <td><%= f.weight %></td>
+                      <td><% if (f.hidden) { %><span class="badge badge-hidden">Hidden</span><% } else { %><span class="badge badge-active">Active</span><% } %></td>
+                    </tr>
+                    <% }) %>
+                  </tbody>
+                </table>
+              </td>
+            </tr>
+            <% } %>
+            <% }) %>
+          </tbody>
+        </table>
+        </div>
+        <div id="no-results" class="empty-msg empty-msg-padded is-hidden">No commands match your filter.</div>
+      </div>
+    </section>
+
+  </main>
+
+  <%- include('partials/pwa-register') %>
+  <script src="/sfx.js"></script>
+</body>
+</html>
